switched rawdata to double

+++ app/src/androidTest/java/onetallprogrammer/hackpoly2018/probability/RawDataTest.java
@@ -16,15 +16,15 @@ public class RawDataTest {
 
     @Before
     public void setup(){
-        rawData.addPoint(1);
-        rawData.addPoint(2);
-        rawData.addPoint(3);
+        rawData.addPoint(1.0);
+        rawData.addPoint(2.0);
+        rawData.addPoint(3.0);
     }
 
     @Test
     public void getMeanTest() throws Exception {
-        float expected = 2.0f;
-        float mean = rawData.getMean();
+        Double expected = 2.0;
+        Double mean = rawData.getMean();
 //        Log.d("mean", String.valueOf(mean));
         boolean inRange = Math.abs(expected - mean) < epsilon;
 
@@ -33,8 +33,8 @@ public class RawDataTest {
 
     @Test
     public void getSumTest() throws Exception{
-        float sum = rawData.getSum();
-        float expected = 6.0f;
+        Double sum = rawData.getSum();
+        Double expected = 6.0;
         boolean inRange = Math.abs(sum - expected) < epsilon;
 
         assertTrue(inRange);
@@ -42,8 +42,8 @@ public class RawDataTest {
 
     @Test
     public void getStdDevTest() throws Exception{
-        float stdDev = rawData.getStdDev();
-        float expected = 0.8165f;
+        Double stdDev = rawData.getStdDev();
+        Double expected = 0.8165;
         boolean inRange = Math.abs(expected - stdDev) < epsilon;
 
         assertTrue(inRange);
